Fix firewalld module invocations (immediate)
A couple of recent tests had incoming HTTP/S traffic not enabled.  The
permanent configuration had been applied to firewalld but the ports
were not opened (so a simple reload fixed).  The setup of several tasks
had "immediate: true" not set but instead relied on a restart of
firewalld being done through a notify at some point.  Switched out the
dependence on a notify for explicit coding of "immediate: true".