Protecting the remotes from later changes
The early remote ELANs get their values overwritten by the values from
the subsequent ELAN.  This is broken.  Making the template and copy
module invocations conditional (force: false) so that they only get
changed if the files don't already exist.  Also, the podman systemd
stuff seems to be hella fragile so making that conditional as well.